Fixes and tests
- Empty dirs were not archived
- File or dir included multiple times, led to multiple
  inclusions in the archive

Found by fuzz testing.

+++ test/test_create_sets.py
@@ -6,11 +6,13 @@ from impl.upload_sets import build_archive, get_list_files
 
 import os
 import pathlib
+import pickle
 import pytest
 import random
 import shutil
 import string
 import subprocess
+import sys
 import time
 
 SCRIPT_PATH = os.path.dirname(os.path.abspath(__file__))
@@ -25,6 +27,8 @@ SNAPSHOT_PATHS = (
     '/glacier_deep_archive_backup_test',
     '/mnt/glacier_deep_archive_backup_test')
 SET_PATH = os.path.join(WORK_PATH, 'sets')
+REPRO_PATH = os.path.join(SCRIPT_PATH, 'state', 'repro.pickle.')
+os.makedirs(os.path.dirname(REPRO_PATH), exist_ok=True)
 
 def create_files(items):
     for item_path, size in items:
@@ -36,7 +40,11 @@ def create_files(items):
             with open(os.path.join(POOL_PATH, item_path), 'wt') as f:
                 f.write('x' * size)
 
-def run_test_for_snapshot_paths(snapshot_path, pool_files, backup_paths, upload_limit): # pylint: disable=too-many-branches
+def run_test_for_snapshot_paths(snapshot_path, pool_files, backup_paths, # pylint: disable=too-many-branches, too-many-statements
+                                upload_limit, num_expected_sets):
+    snapshot_path = os.path.normpath(snapshot_path)
+    backup_paths = tuple(map(os.path.normpath, backup_paths))
+
     for item in os.listdir(WORK_PATH):
         shutil.rmtree(os.path.join(WORK_PATH, item), ignore_errors=True)
     os.makedirs(POOL_PATH)
@@ -56,6 +64,9 @@ def run_test_for_snapshot_paths(snapshot_path, pool_files, backup_paths, upload_
             return
         raise Exception('backup_paths not empty, but no set files found')
 
+    if num_expected_sets is not None and len(list_files) != num_expected_sets:
+        raise Exception(f"Wrong number of sets={len(list_files)}, expected={num_expected_sets}")
+
     archive_paths = []
     for list_file in list_files:
         _, archive_path = build_archive(list_file, WORK_PATH)
@@ -65,6 +76,7 @@ def run_test_for_snapshot_paths(snapshot_path, pool_files, backup_paths, upload_
     os.makedirs(extract_path, exist_ok=True)
 
     for archive_path in archive_paths:
+        print(f"Extracting {archive_path}")
         cmd = (os.path.join(SCRIPT_PATH, '..', 'extract_archive'), archive_path, extract_path)
         subprocess.run(cmd, check=True)
 
@@ -72,12 +84,18 @@ def run_test_for_snapshot_paths(snapshot_path, pool_files, backup_paths, upload_
     extract_base_path = os.path.join(extract_path, relative_snapshot_path)
 
     # Condition 1: All backup paths must be identical
+    extract_backup_paths = set()
     for backup_path in backup_paths:
         snapshot_backup_path = os.path.join(snapshot_path, backup_path)
         extract_backup_path = os.path.join(extract_base_path, backup_path)
         cmd = ('diff', '-r', snapshot_backup_path, extract_backup_path)
-        print(cmd)
+        print(' '.join(cmd))
         subprocess.run(cmd, check=True)
+        extract_backup_paths.add(extract_backup_path)
+
+    for extract_backup_path in extract_backup_paths:
+        if not os.path.exists(extract_backup_path):
+            continue
         if os.path.islink(extract_backup_path) or os.path.isfile(extract_backup_path):
             os.remove(extract_backup_path)
         else:
@@ -108,9 +126,24 @@ def run_test_for_snapshot_paths(snapshot_path, pool_files, backup_paths, upload_
         raise Exception(f"Extract dir {extract_base_path} has extraneous items:"
                         f" files={extra_files}, dirs={extra_dirs}")
 
-def run_test(pool_files, backup_paths, upload_limit):
+def run_test(pool_files, backup_paths, upload_limit, num_expected_sets=None, is_fuzz_run=False):
     for snapshot_path in SNAPSHOT_PATHS:
-        run_test_for_snapshot_paths(snapshot_path, pool_files, backup_paths, upload_limit)
+        try:
+            run_test_for_snapshot_paths(snapshot_path, pool_files, backup_paths,
+                                        upload_limit, num_expected_sets)
+        except:
+            if is_fuzz_run:
+                i = 0
+                while 1:
+                    repro_path = f"{REPRO_PATH}{i}"
+                    i += 1
+                    if not os.path.exists(repro_path):
+                        break
+                with open(repro_path, 'wb') as f:
+                    pickle.dump((snapshot_path, pool_files, backup_paths, upload_limit), f)
+                print(f"Wrote reproducer to '{repro_path}', run"
+                      f" '{__file__} --repro {repro_path}' to reproduce")
+            raise
 
 # Letters: directories, numbers: files
 # Trailing slash means directory
@@ -126,7 +159,7 @@ def test_one_set_single_file():
         ('a/1', SIZE_SMALL),
     )
 
-    run_test(pool_files, ('a', ), SIZE_SMALL)
+    run_test(pool_files, ('a', ), SIZE_SMALL, num_expected_sets=1)
 
 def test_one_set_multiple_files():
     pool_files = (
@@ -135,7 +168,7 @@ def test_one_set_multiple_files():
         ('a/b/3', SIZE_SMALL)
     )
 
-    run_test(pool_files, ('a', ), SIZE_SMALL * 3)
+    run_test(pool_files, ('a', ), SIZE_SMALL * 3, num_expected_sets=1)
 
 def test_three_dirs_include_two():
     pool_files = (
@@ -144,7 +177,7 @@ def test_three_dirs_include_two():
         ('c/3', SIZE_SMALL),
     )
 
-    run_test(pool_files, ('a', 'c'), SIZE_SMALL * 3)
+    run_test(pool_files, ('a', 'c'), SIZE_SMALL * 3, num_expected_sets=1)
 
 def test_file_and_subdir():
     pool_files = (
@@ -152,7 +185,7 @@ def test_file_and_subdir():
         ('a/b/2', SIZE_SMALL),
     )
 
-    run_test(pool_files, ('a', ), SIZE_SMALL)
+    run_test(pool_files, ('a', ), SIZE_SMALL, num_expected_sets=2)
 
 def test_top_level_files_one_set():
     pool_files = (
@@ -160,7 +193,7 @@ def test_top_level_files_one_set():
         ('2', SIZE_SMALL),
     )
 
-    run_test(pool_files, ('1', '2'), SIZE_SMALL*2)
+    run_test(pool_files, ('1', '2'), SIZE_SMALL*2, num_expected_sets=1)
 
 def test_top_level_files_two_sets():
     pool_files = (
@@ -168,7 +201,7 @@ def test_top_level_files_two_sets():
         ('2', SIZE_SMALL),
     )
 
-    run_test(pool_files, ('1', '2'), SIZE_SMALL)
+    run_test(pool_files, ('1', '2'), SIZE_SMALL, num_expected_sets=2)
 
 def test_top_level_files_and_dirs():
     pool_files = (
@@ -180,12 +213,12 @@ def test_top_level_files_and_dirs():
         ('d/e/f/g/6', SIZE_SMALL),
     )
 
-    run_test(pool_files, ('1', '2', 'a', 'c', 'd'), SIZE_SMALL * 2)
-    run_test(pool_files, ('1', '2', 'a', 'c', 'd'), SIZE_SMALL)
-    run_test(pool_files, ('1', 'a', 'c', 'd'), SIZE_SMALL * 2)
-    run_test(pool_files, ('1', 'a', 'd'), SIZE_SMALL * 2)
-    run_test(pool_files, ('2', 'a', 'd'), SIZE_SMALL * 2)
-    run_test(pool_files, ('1', 'a', 'c'), SIZE_SMALL * 2)
+    run_test(pool_files, ('1', '2', 'a', 'c', 'd'), SIZE_SMALL * 2, num_expected_sets=3)
+    run_test(pool_files, ('1', '2', 'a', 'c', 'd'), SIZE_SMALL, num_expected_sets=6)
+    run_test(pool_files, ('1', 'a', 'c', 'd'), SIZE_SMALL * 2, num_expected_sets=3)
+    run_test(pool_files, ('1', 'a', 'd'), SIZE_SMALL * 2, num_expected_sets=2)
+    run_test(pool_files, ('2', 'a', 'd'), SIZE_SMALL * 2, num_expected_sets=2)
+    run_test(pool_files, ('1', 'a', 'c'), SIZE_SMALL * 2, num_expected_sets=2)
 
 def test_file_size_exceeds_upload_limit_throws():
     pool_files = (
@@ -204,7 +237,7 @@ def test_path_fits_but_only_one_file_included():
         ('a/b/2', SIZE_SMALL),
     )
 
-    run_test(pool_files, ('a/b/1',), SIZE_SMALL * 2)
+    run_test(pool_files, ('a/b/1',), SIZE_SMALL * 2, num_expected_sets=1)
 
 def test_path_fits_but_only_one_subdir_included():
     pool_files = (
@@ -212,7 +245,35 @@ def test_path_fits_but_only_one_subdir_included():
         ('a/c/2', SIZE_SMALL),
     )
 
-    run_test(pool_files, ('a/b',), SIZE_SMALL * 2)
+    run_test(pool_files, ('a/b',), SIZE_SMALL * 2, num_expected_sets=1)
+
+def test_empty_dir():
+    pool_files = (
+        ('a/', 0),
+    )
+
+    run_test(pool_files, ('a',), SIZE_SMALL, num_expected_sets=1)
+
+def test_dir_specified_multiple_times():
+    pool_files = (
+        ('a/1', SIZE_SMALL),
+    )
+
+    run_test(pool_files, ('a', 'a'), SIZE_SMALL, num_expected_sets=1)
+
+def test_file_specified_multiple_times():
+    pool_files = (
+        ('1', SIZE_SMALL),
+    )
+
+    run_test(pool_files, ('1', '1'), SIZE_SMALL, num_expected_sets=1)
+
+def test_dir_trailing_slash():
+    pool_files = (
+        ('a/', 0),
+    )
+
+    run_test(pool_files, ('a/',), SIZE_SMALL, num_expected_sets=1)
 
 def do_test_fuzz():
     MAX_FILES = 1000
@@ -251,12 +312,7 @@ def do_test_fuzz():
 
     upload_limit = random.randint(MAX_FILE_SIZE, 3 * MAX_FILE_SIZE)
 
-    try:
-        run_test(pool_files, backup_paths, upload_limit)
-    except:
-        print(f"test_fuzz failed, inputs: pool_files={pool_files}, "
-              f"backup_paths={backup_paths}, upload_limit={upload_limit}")
-        raise
+    run_test(pool_files, backup_paths, upload_limit, is_fuzz_run=True)
 
 @pytest.mark.longrunner
 def test_fuzz():
@@ -270,12 +326,20 @@ def test_fuzz():
     return NUM_RUNS
 
 if __name__ == '__main__':
-    # Endless fuzzing
-    total_runs = 0
-    try:
-        while 1:
-            total_runs += test_fuzz()
-    finally:
-        print(f"Runs: {total_runs}")
+    if '--repro' in sys.argv:
+        repro_path = sys.argv[sys.argv.index('--repro') + 1]
+        with open(repro_path, 'rb') as f:
+            args = pickle.load(f)
+            run_test_for_snapshot_paths(*args, num_expected_sets=None)
+    elif '--fuzz' in sys.argv:
+        # Endless fuzzing
+        total_runs = 0
+        try:
+            while 1:
+                total_runs += test_fuzz()
+        finally:
+            print(f"Runs: {total_runs}")
+    else:
+        print('Valid args: --repro, --fuzz')
 
 subprocess.run(('sudo', 'umount', WORK_PATH), check=True)
